+++ src/modules/statisticDashboard/tests/unit/components/StatisticDashboard.spec.js
@@ -234,6 +234,9 @@ describe("src/modules/StatisticDashboard.vue", () => {
                 }),
                 spyCheckFilterSettings = sinon.stub(wrapper.vm, "checkFilterSettings");
 
+            wrapper.vm.setFlattenedRegions([{
+                values: ["foo", "bar"]
+            }]);
             wrapper.vm.setSelectedReferenceData("too");
             wrapper.vm.setSelectedDates(["too"]);
             wrapper.vm.setSelectedRegions(["too"]);
@@ -672,7 +675,6 @@ describe("src/modules/StatisticDashboard.vue", () => {
 
                 wrapper.vm.getFilter(regions, dates);
                 expect(getFilterForListSpy.getCall(0).args).to.deep.equal([dates, "date"]);
-                expect(getFilterForListSpy.getCall(1).args).to.deep.equal([["bar"], "bar"]);
                 sinon.restore();
             });
         });